revise and shorten source code
